backend/requirements.txt
flask==2.0.1
Flask-CORS
Pillow